Add logging in controllers.

+++ MVCNBlog/Controllers/AccountController.cs
@@ -5,6 +5,7 @@ using System.Web;
 using System.Web.Mvc;
 using System.Web.Security;
 using BLL.Interface.Services;
+using LoggingModule;
 using MVCNBlog.Infrastructure.Mappers;
 using MVCNBlog.Providers;
 using MVCNBlog.ViewModels;
@@ -16,9 +17,11 @@ namespace MVCNBlog.Controllers
     public class AccountController : Controller
     {
         private readonly IAccountService service;
+        private readonly ILogger logger;
 
-        public AccountController(IAccountService service)
+        public AccountController(IAccountService service, ILogger logger)
         {
+            this.logger = logger;
             this.service = service;
         }
 
@@ -47,15 +50,12 @@ namespace MVCNBlog.Controllers
                     {
                         return Redirect(returnUrl);
                     }
-                    else
-                    {
-                        return RedirectToAction("All", "Article");
-                    }
+
+                    return RedirectToAction("All", "Article");
                 }
-                else
-                {
-                    ModelState.AddModelError("", "Incorrect login or password.");
-                }
+
+                logger.Info($"Can not login with name: {viewModel.Name}");
+                ModelState.AddModelError("", "Incorrect login or password.");
             }
             return View(viewModel);
         }
@@ -88,12 +88,10 @@ namespace MVCNBlog.Controllers
                 if (membershipUser != null)
                 {
                     FormsAuthentication.SetAuthCookie(viewModel.Name, false);
-                    return RedirectToAction("Index", "Home");
+                    return RedirectToAction("All", "Article");
                 }
-                else
-                {
-                    ModelState.AddModelError("", "Error registration.");
-                }
+                logger.Info($"Can not register with name: {viewModel.Name}");
+                ModelState.AddModelError("", "Error registration.");
             }
             return View(viewModel);
         }
@@ -112,8 +110,12 @@ namespace MVCNBlog.Controllers
         {
             var currentAccount = service.GetAccountEntity(User.Identity.Name).ToMvcAccount();
             if (currentAccount == null)
-                return HttpNotFound("NotFound.");
-            
+            {
+                var httpException = new HttpException(404, "Not found");
+                logger.Warn(httpException, $"{nameof(currentAccount)} wasnt found.");
+                throw httpException;
+            }
+
             return View(currentAccount);
         }
 
@@ -123,8 +125,12 @@ namespace MVCNBlog.Controllers
         public ActionResult Edit()
         {
             var currentAccount = service.GetAccountEntity(User.Identity.Name).ToMvcAccount();
-            if (currentAccount == null)
-                return HttpNotFound("NotFound.");
+            if (currentAccount == null) if (currentAccount == null)
+                {
+                    var httpException = new HttpException(404, "Not found");
+                    logger.Warn(httpException, $"{nameof(currentAccount)} wasnt found.");
+                    throw httpException;
+                }
 
             return View(currentAccount);
         }
@@ -153,7 +159,11 @@ namespace MVCNBlog.Controllers
         {
             var currentAccount = service.GetAccountEntity(User.Identity.Name).ToMvcAccount();
             if (currentAccount == null)
-                return HttpNotFound("NotFound.");
+            {
+                var httpException = new HttpException(404, "Not found");
+                logger.Warn(httpException, $"{nameof(currentAccount)} wasnt found.");
+                throw httpException;
+            }
 
             return View(currentAccount);
         }
@@ -163,6 +173,13 @@ namespace MVCNBlog.Controllers
         public ActionResult ConfirmDelete()
         {
             var currentAccount = service.GetAccountEntity(User.Identity.Name);
+            if (currentAccount == null)
+            {
+                var httpException = new HttpException(404, "Not found");
+                logger.Warn(httpException, $"{nameof(currentAccount)} wasnt found.");
+                throw httpException;
+            }
+
             service.DeleteAccount(currentAccount);
 
             return RedirectToAction("LogOff");
